Move showStats function to utils
